Fixed including empty filters in the query string
When a filter was added with a NULL value, then the filter key was not
added to the query string. This is now fixed.
